[atServer] Fix command length > max_length
In case the user enters a command string which
length is higher than atServer's max command length
return ERROR

Resolves: LE-5737

+++ components/atServices/atClient/le_atClient.c
@@ -606,7 +606,7 @@ static void RxNewData
         LE_CRIT("Unexpected event(s) on fd %d (0x%hX).", fd, events);
     }
 
-    int32_t size = 0;
+    ssize_t size = 0;
     DeviceContext_t *interfacePtr = le_fdMonitor_GetContextPtr();
 
     LE_DEBUG("Start read");
